Move server code into different files
