+++ README.md
@@ -25,22 +25,25 @@ This is mostly for my own benefit, but I'll briefly list the files used.
 * **conf.lua**: Sets up important stuff like window size. It is called by LOVE at the beginning.
 * **game-setup.lua**: Contains functions for receiving game data from server and initiating the game on the client.
 * **game-updates**: Contains code for receiving game updates from server in any of the four game modes.
-* **gameDeckPlay.lua**: Contains the gamestate for the flip part of the game, when it is your go.
-* **gameDeckWait.lua**: Contains the gamestate for the flip part of the game, when it is not your go.
-* **gameHandPlay.lua**: Contains the gamestate for the hand part of the game, when it is your go.
-* **gameHandWait.lua**: Contains the gamestate for the hand part of the game, when it is not your go.
-* **gameOver.lua**: Contains the gamestate for when the game has finished, either when someone stops continuing or cards run out.
+* **game-deck-play.lua**: Contains the gamestate for the flip part of the game, when it is your go.
+* **game-deck-wait.lua**: Contains the gamestate for the flip part of the game, when it is not your go.
+* **game-hand-play.lua**: Contains the gamestate for the hand part of the game, when it is your go.
+* **game-hand-wait.lua**: Contains the gamestate for the hand part of the game, when it is not your go.
+* **gameover.lua**: Contains the gamestate for when the game has finished, either when someone stops continuing or cards run out.
 * **main.lua**: The main client file. The callbacks set here are called by LOVE when things happen. Mostly calls functions of whatever gamestate is currently happening and moves to the next gamestate.
 * **main-old.lua**: The old main client file, not used anymore. Included mostly for entertainment, I guess?
 * **menu.lua**: Contains a gamestate that controls logging in at the beginning of the game.
-* **theyScore.lua**: Contains a gamestate for when the other player scores.
+* **they-score.lua**: Contains a gamestate for when the other player scores.
 * **tween.lua**: Used for moving cards around smoothly.
 * **waiting.lua**: Contains a gamestate for waiting for another player to arrive.
-* **youScore.lua**: Contains a gamestate for when you score and decide whether to continue.
+* **you-score.lua**: Contains a gamestate for when you score and decide whether to continue.
 
 #### Server
-* **server.lua**: It controls the server and gets it to send and receive messages.
 * **cards-score.lua**: It takes a set of cards collected and scores them according to the rules of Koi-Koi.
+* **game-end-updates.lua**: Controls the game near the end, with sending and receiving Koi-Koi updates and sending the score when the game ends.
+* **game-updates.lua**: Controls normal play updates in all modes and sends out messages.
+* **server.lua**: It controls the server and gets it to send and receive messages.
+* **set-up-game.lua**: Contains code for setting up a new game and sending it out.
 
 #### Both
 * **cards-define.lua**: Defines all cards, their values, months and images.
